Adds traversal and pruning.
This is used to remove 'global cycles' after the tree has been
constructed. A 'global cycle' is a sequence of arcs where one or more
arcs appear more than once.

In the future, this might need to be re-coded in order to allow for
other test and action functions during traversal.

This may be used not only as a final pruning step, but also to reduce
the graph's growth during construction.

git-svn-id: https://ad-svn.informatik.uni-freiburg.de/forestmaps@43 6bfcf819-cacf-47a1-ae15-21b88a40d8fc