Changes the way parking lots are taken into account
After computing the popularity for regular forest entries and
distributing the car population over parking lots (taking into account
their ranking) the parking lots are regarded as forest entries in the
second phase.
To not change the code too much the "new" forest entries are added to
the forest entry description file at the end of phase 1) reachability.

git-svn-id: https://ad-svn.informatik.uni-freiburg.de/forestmaps@181 6bfcf819-cacf-47a1-ae15-21b88a40d8fc